stealthpay-backend-main/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
sqlalchemy
aiosqlite
plaid-python
stripe
web3
httpx
python-multipart
alembic
